+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/FakeTrackSelection.java
@@ -25,6 +25,7 @@ import androidx.media3.common.util.UnstableApi;
 import androidx.media3.exoplayer.source.chunk.MediaChunk;
 import androidx.media3.exoplayer.source.chunk.MediaChunkIterator;
 import androidx.media3.exoplayer.trackselection.ExoTrackSelection;
+import com.google.common.base.Objects;
 import java.util.List;
 
 /**
@@ -35,13 +36,19 @@ import java.util.List;
 public final class FakeTrackSelection implements ExoTrackSelection {
 
   private final TrackGroup rendererTrackGroup;
+  private final int selectedIndex;
 
   public int enableCount;
   public int releaseCount;
   public boolean isEnabled;
 
   public FakeTrackSelection(TrackGroup rendererTrackGroup) {
+    this(rendererTrackGroup, /* selectedIndex= */ 0);
+  }
+
+  public FakeTrackSelection(TrackGroup rendererTrackGroup, int selectedIndex) {
     this.rendererTrackGroup = rendererTrackGroup;
+    this.selectedIndex = selectedIndex;
   }
 
   // TrackSelection implementation.
@@ -68,18 +75,23 @@ public final class FakeTrackSelection implements ExoTrackSelection {
 
   @Override
   public int getIndexInTrackGroup(int index) {
-    return 0;
+    return index;
   }
 
   @Override
   public int indexOf(Format format) {
     assertThat(isEnabled).isTrue();
-    return 0;
+    for (int i = 0; i < rendererTrackGroup.length; i++) {
+      if (rendererTrackGroup.getFormat(i).equals(format)) {
+        return i;
+      }
+    }
+    return -1;
   }
 
   @Override
   public int indexOf(int indexInTrackGroup) {
-    return 0;
+    return indexInTrackGroup;
   }
 
   // ExoTrackSelection specific methods.
@@ -102,17 +114,17 @@ public final class FakeTrackSelection implements ExoTrackSelection {
 
   @Override
   public Format getSelectedFormat() {
-    return rendererTrackGroup.getFormat(0);
+    return rendererTrackGroup.getFormat(selectedIndex);
   }
 
   @Override
   public int getSelectedIndexInTrackGroup() {
-    return 0;
+    return selectedIndex;
   }
 
   @Override
   public int getSelectedIndex() {
-    return 0;
+    return selectedIndex;
   }
 
   @Override
@@ -157,5 +169,27 @@ public final class FakeTrackSelection implements ExoTrackSelection {
   public boolean isTrackExcluded(int index, long nowMs) {
     assertThat(isEnabled).isTrue();
     return false;
+  }
+
+  @Override
+  public boolean equals(@Nullable Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (!(o instanceof FakeTrackSelection)) {
+      return false;
+    }
+    FakeTrackSelection that = (FakeTrackSelection) o;
+    return enableCount == that.enableCount
+        && releaseCount == that.releaseCount
+        && isEnabled == that.isEnabled
+        && selectedIndex == that.selectedIndex
+        && Objects.equal(rendererTrackGroup, that.rendererTrackGroup);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hashCode(
+        rendererTrackGroup, enableCount, releaseCount, isEnabled, selectedIndex);
   }
 }